Refactor code, add nodemon, fix Redis connection, and clean up submission logic
