+++ .env.example
@@ -1,0 +1,5 @@
+PORT=3000
+MONGODB_URL=mongodb://localhost:27017/codejudge
+JWT_SECRET=your_jwt_secret_here
+JWT_EXPIRY=1d
+REDIS_URL=redis://localhost:6379
